InferenceTimeTreeSearch/pyproject.toml
[project]
name = "knowledge-agent"
version = "0.1.0"
description = "LangGraph를 사용한 지식 기반 질의응답 에이전트"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]
